Automatically set CookieSecure from request, add CookieSameSiteHelper helper function
Previously this was set at startup depending on which flags were used,
which was rather annoying and could lead to confusing situations. Most
proxies seem to set X-Forwarded-Proto in my testing, so we can use that.

Also replace the CookieSameSite global with a CookieSameSiteHelper(), so
it can be set per-request.